runny-refactor: massive refactor to use Runny for processes
Runny is a new, external crate that takes care of process running.

This refactor is responsible for cleaning up much of the codebase
to take advantage of this external package.

As part of this refactor, we also remove many of the big lock
around the main TestSet, which prepares it for dynamic tests in
the future.

Signed-off-by: Sean Cross <sean@xobs.io>